Add a Datadog client that pages people upon events
[#50018413]

This is implemented as a decorator object to limit the leaky abstraction
that Datadog forces upon us. At the moment the health monitor is
required to know that we are using PagerDuty for our alerting. In the
future this change should be easy to remove so we can use the vanilla
Datadog client again.